requirements.txt
discord.py
yt_dlp
asyncio
PyNaCl
pydub
ffmpeg
ytmusicapi
google-auth-oauthlib
google-api-python-client
pytube
